move scroll checks to sidebar
