+++ src/app/report-rollup/report-rollup.component.html
@@ -58,7 +58,7 @@
 
     <div class="d-flex">
 
-        <app-sidebar [sidebarHeight]="sidebarHeight" [bannerHeight]="bannerHeight" [focusedAssessment]="focusedAssessment"></app-sidebar>
+        <app-sidebar [sidebarHeight]="sidebarHeight" [bannerHeight]="bannerHeight"></app-sidebar>
         <!-- End .sidebar -->
 
         <div #assessmentReportsDiv class="assessment-padding d-flex flex-column scroll-item print-height w-100 skyline">
